pyproject.toml
[tool.poetry]
name = "vtuber-twitter-tools"
version = "0.1.0"
description = ""
authors = ["新篠ゆう <37145593+yuarasino@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "vtuber_twitter_tools"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pillow = "^9.2.0"
tzdata = "^2022.5"
requests-oauthlib = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
